Fix for MTU issue.
This allows bigger Econet MTU of 8192 bytes.

There are some other issues with big frames that I need to work through
related to timing.  Essentially, I think we're going to need to bring the
PARLIO driver in tree to get the control we need in order to maximise the
DMA usage.